+++ src/ts/Ana/Ana.ts
@@ -3,6 +3,7 @@ import { getElements } from '../Elements/Elements'
 import { RenderDictionary } from '../types'
 import { AttributeValuesDictionary, MatchFunctionsDictionary } from '../types'
 import * as utils from '../utils'
+import { rLayout } from '../Components/Atoms/Layout/Layout'
 // import { rDisplay } from '../Components/Atoms/Display/Display'
 // import { rTitle } from '../Components/Atoms/Title/Title'
 // import { rTestpage } from '../Components/Organisms/Testpage/Testpage'
@@ -92,15 +93,16 @@ export class Ana implements iAna {
     HTMLElement.prototype.has = this.has
     HTMLElement.prototype.setAttributes = this.setAttributes
     let elements = getElements(this.configuration)
-    // let atoms = {
-    //   Display: rDisplay(elements, this.configuration),
-    //   Title: rTitle(elements, this.configuration),
-    // }
+    let atoms = {
+      Layout: rLayout(elements, this.configuration)
+      //   Display: rDisplay(elements, this.configuration),
+      //   Title: rTitle(elements, this.configuration),
+    }
     // let organisms = {
     //   Testpage: rTestpage(elements, this.configuration)
     // }
     // return { ...elements, ...atoms, ...organisms }
-    return { ...elements }
+    return { ...elements, ...atoms}
   }
 
   /**